changed types of alpha in elu

+++ src/computation/op/def2/elu.cpp
@@ -19,7 +19,7 @@ namespace mv
                 return {false, 0};
             }
             
-            auto alpha = args.at("alpha").get<double>();
+            auto alpha = args.at("alpha").get<unsigned>();
 
             if (alpha < 0)
             {
@@ -42,7 +42,7 @@ namespace mv
     
         MV_REGISTER_OP(Elu)
         .setInputs({"data"})
-        .setArg<double>("alpha") 
+        .setArg<unsigned>("alpha") 
         .setOutputs({"output"})
         .setInputCheck(inputCheckFcn)
         .setOutputDef(outputDefFcn)
